% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccvx5_build_jags.R
\name{ccvx5_build_jags}
\alias{ccvx5_build_jags}
\title{Build 5-parameter Concavex JAGS code}
\usage{
ccvx5_build_jags(prior.theta_0 = "theta_0 ~ dnorm(0, 1E-4)",
  prior.theta_1 = "theta_1 ~ dnorm(0, 1E-4)",
  prior.gamma_1 = "gamma_1 ~ dunif(.001, .999)",
  prior.gamma_2 = "gamma_2 ~ dunif(.001, .999)",
  prior.alpha = "alpha ~ dunif(0.001, 3.1415926/2 - .001)",
  predictive.probs = FALSE)
}
\arguments{
\item{prior.theta_0}{Prior for theta_0 specified in terms of JAGS code. The default is a non-informative normal prior: 'theta_0 ~ dnorm(0, 1E-4)'}

\item{prior.theta_1}{Prior for theta_1 specified in terms of JAGS code. The default is a non-informative normal prior: 'theta_1 ~ dnorm(0, 1E-4)'}

\item{prior.gamma_1}{Prior for gamma_1 specified in terms of JAGS code. The default is a non-informative uniform prior: 'gamma_1 ~ dunif(-.001, .999)'}

\item{prior.gamma_2}{Prior for gamma_2 specified in terms of JAGS code. The default is a non-informative uniform prior: 'gamma_2 ~ dunif(-.001, .999)'}

\item{prior.alpha}{Prior for alpha specified in terms of JAGS code. The default is a non-informative uniform prior on (0, pi/2): 'alpha ~ dunif(0.001, 3.1415926/2 - .001)'}

\item{predictive.probs}{Indicator for whether to compute posterior predictive probabilites for a future study. default is FALSE. if set to TRUE, the ccvx_fit function requires additional inputs}
}
\description{
This function creates JAGS code within R for fitting the 5 parameter Concavex model. It has default non-informative priors for model parameters, but allows for specification of custom priors written in JAGS syntax
}
\examples{
ccvx.jags.mod <- ccvx5_build_jags()
cat(ccvx.jags.mod)
}
