% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccvx_build_jags.R
\name{ccvx_build_jags}
\alias{ccvx_build_jags}
\title{Build Concavex JAGS code}
\usage{
ccvx_build_jags()
}
\description{
This function creates the JAGS code within R for fitting the Concavex model. Will eventually allow for flexible specification of priors
}
\examples{
ccvx.jags.mod <- ccvx_build_jags()
cat(ccvx.jags.mod)
}
