% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccvx_ddcp_plot.R
\name{ccvx_ddcp_plot}
\alias{ccvx_ddcp_plot}
\title{Plot Concavex model posterior predictive risks for Phase 3 (AKA data-driven conditional probabilities)}
\usage{
ccvx_ddcp_plot(ccvx.samples, eff.thresholds = NULL)
}
\arguments{
\item{ccvx.samples}{List containg Gibbs samples output by ccvx_fit() function}

\item{eff.thresholds}{A list of efficacy thresholds of specific interest for which to compute probabilities}
}
\description{
Function to plot Concavex model risk profiles. Specifically, posterior predictive probabilities of exceeding efficacy thresholds
are plotted against doses tested, and against specific thresholds of interest. If no thresholds are specified in 
'eff.thresholds' argument, only the plot against dose is generated.
}
\examples{
ccvx.mod <- ccvx_build_jags(predictive.probs = TRUE)
ccvx.samples <- ccvx_fit(ccvx.mod, doses = 0:4, mu.hat = c(1, 20, 50, 60, 65), std.err = rep(20, 5), sd.ph3 = 200, n.per.arm.ph3 = 300)
par(mfrow=c(1,2))
ccvx_ddcp_plot(ccvx.samples, eff.thresholds = c(-5, 0, 5, 10, 20, 40))

ccvx.mod <- ccvx5_build_jags(predictive.probs = TRUE)
ccvx.samples <- ccvx_fit(ccvx.mod, doses = 0:4, mu.hat = c(1, 20, 50, 60, 65), std.err = rep(20, 5), sd.ph3 = 200, n.per.arm.ph3 = 300)
par(mfrow=c(1,2))
ccvx_ddcp_plot(ccvx.samples, eff.thresholds = c(-5, 0, 5, 10, 20, 40))
}
