% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccvx_fit.R
\name{ccvx_fit}
\alias{ccvx_fit}
\title{Sample from Concavex model posteriors}
\usage{
ccvx_fit(ccvx.mod, doses, mu.hat, stderr, n.chains = 4,
  gibbs.samples = 5000)
}
\arguments{
\item{ccvx.mod}{JAGS model file as specified by ccvx_build_jags()}

\item{doses}{A vector of dose strengths}

\item{mu.hat}{A vector of parameter estimates for each dose}

\item{stderr}{A vector of standard errors for parameter estimates}

\item{n.chains}{Number of chains used to sample. Default is 4}

\item{gibbs.samples}{Number of samples to draw for each chain. Default is 5000}
}
\description{
This function samples from posterior densities of concavex model parameters as well as transformations of these parameters.
}
\examples{
ccvx.mod <- ccvx_build_jags()
ccvx.samples <- ccvx_fit(ccvx.mod, doses = 0:4, mu.hat = c(1, 20, 50, 60, 65), stderr = rep(10, 5))
ccvx_plot_fit(ccvx.samples)
}
