% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccvx_hist_post.R
\name{ccvx_hist_post}
\alias{ccvx_hist_post}
\title{Plot Concavex model fit}
\usage{
ccvx_hist_post(ccvx.samples, cred.int.width = 0.9)
}
\arguments{
\item{ccvx.samples}{List containg Gibbs samples output by ccvx_fit() function}

\item{cred.int.width}{Specify width of credible interval for posteriors}
}
\description{
This function plots the Concavex model fit overlaid with parameter estimates and standard errors for each dose
}
\examples{
ccvx.mod <- ccvx_build_jags()
ccvx.samples <- ccvx_fit(ccvx.mod, doses = 0:4, mu.hat = c(1, 20, 50, 60, 65), stderr = rep(10, 5))
par(mfrow=c(1,3))
ccvx_hist_post(ccvx.samples)
}
