% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccvx_plot_fit.R
\name{ccvx_plot_fit}
\alias{ccvx_plot_fit}
\title{Plot Concavex model fit}
\usage{
ccvx_plot_fit(ccvx.samples, placebo.adjusted = FALSE,
  title = "Concavex Model Fit", xlab = "Dose", ylab = "Effect",
  cred.int.width = 0.9)
}
\arguments{
\item{ccvx.samples}{List containg Gibbs samples output by ccvx_fit() function}

\item{placebo.adjusted}{Should the estimated dose-response curve be placebo adjusted? Default is FALSE}

\item{title}{Plot title}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{cred.int.width}{Specify width of credible interval for curve. This is also applied to confidence intervals of point estimates. Default is 0.9 (90\%)}
}
\description{
Function to plot Concavex model fit overlaid on actual points
}
\examples{
ccvx.mod <- ccvx_build_jags()
ccvx.samples <- ccvx_fit(ccvx.mod, doses = 0:4, mu.hat = c(1, 20, 50, 60, 65), std.err = rep(10, 5))
par(mfrow=c(1,2))
ccvx_plot_fit(ccvx.samples, placebo.adjusted = FALSE, ylab = "Treatment Effect")
ccvx_plot_fit(ccvx.samples, placebo.adjusted = TRUE, ylab = "Pbo-Adjusted Treatment Effect")
}
\author{
Paul Manser, \email{manser.paul@gene.com}
}
