% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccvx_risk_profile.R
\name{ccvx_risk_profile}
\alias{ccvx_risk_profile}
\title{Plot Concavex model risk profiles}
\usage{
ccvx_risk_profile(ccvx.samples, eff.thresholds = NULL)
}
\arguments{
\item{ccvx.samples}{List containg Gibbs samples output by ccvx_fit() function}

\item{eff.thresholds}{A list of efficacy thresholds of specific interest for which to compute probabilities}
}
\description{
Function to plot Concavex model risk profiles. Specifically, posterior probabilities of drug efficacy not exceeding 
specific thresholds are plotted by doses tested, and for specific thresholds of interest. If no thresholds are specified in 
'eff.thresholds' argument, only the plot of doses against treatment effect is generated.
}
\examples{
ccvx.mod <- ccvx_build_jags()
ccvx.samples <- ccvx_fit(ccvx.mod, doses = c(0, 4, 7, 13, 20), mu.hat = c(1, 20, 50, 60, 65), std.err = rep(20, 5))
par(mfrow=c(1,2))
ccvx_risk_profile(ccvx.samples, eff.thresholds = c(5, 10, 20, 40))
}
