% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concavex.R
\docType{package}
\name{concavex}
\alias{concavex}
\alias{concavex-package}
\title{concavex: A package for Bayesian dose-response modeling}
\description{
Provide even more details here!
}
\section{concavex functions}{

The foo functions ...
}

