% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_doc.R
\name{open_doc}
\alias{open_doc}
\title{Access package documents}
\usage{
open_doc(doc = "User Guide", verbose = FALSE)
}
\arguments{
\item{doc}{a character vector specifying the document to open;
see \code{\link{available_docs}}}

\item{verbose}{a logical vector indicating whether the path of the 
document should be printed (Default = FALSE)}
}
\description{
This function allows the user the get access to pdf documents associated
with this package.
}
\author{
Kwame Okrah, \email{okrah.kwame@gene.com}
}
