% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccvx_fit.R
\name{ccvx_fit}
\alias{ccvx_fit}
\title{Sample from Concavex model posteriors}
\usage{
ccvx_fit(ccvx.mod, doses, mu.hat, std.err, n.chains = 4,
  gibbs.samples = 5000, burn.in = 1000, sd.ph3 = NULL,
  n.per.arm.ph3 = NULL)
}
\arguments{
\item{ccvx.mod}{JAGS model file as specified by \link{ccvx_build_jags}}

\item{doses}{A vector of dose strengths, with placebo listed first}

\item{mu.hat}{A vector of parameter estimates for each of the doses}

\item{std.err}{A vector of standard errors for each of the doses}

\item{n.chains}{Number of chains used to for Gibbs sampling. Default is 4}

\item{gibbs.samples}{Number of samples to draw for each chain after burn in. Default is 5000}

\item{burn.in}{Number of burn in samples to be drawn. Default is 1000}

\item{sd.ph3}{The standard deviation of the endpoint being assessed in phase 3. Only needs to be specified when ccvx_fit() computes posterior predictive probabilities (i.e. model-based DDCPs)}

\item{n.per.arm.ph3}{The number of patients per arm being assessed in hypothetical phase 3 study. Only needs to be specified when ccvx_fit() computes posterior predictive probabilities (i.e. model-based DDCPs)}
}
\value{
A list with the elements
\item{ccvx.mod}{The JAGS code used for Gibbs sampling as generated by \link{ccvx_build_jags} or \link{ccvx5_build_jags}}
\item{jags.samples}{output containing Gibbs samples from parameter posteriors generated}
\item{coda.samples}{output containing coda samples for MCMC diagnostics}
\item{doses}{A vector of dose strengths, with placebo listed first}
\item{mu.hat}{A vector of parameter estimates for each of the doses}
\item{std.err}{A vector of standard errors for each of the doses}
}
\description{
This function samples from posterior densities of concavex model parameters as well as transformations of parameters.
}
\examples{
ccvx.mod <- ccvx_build_jags()
ccvx.samples <- ccvx_fit(ccvx.mod, doses = 0:4, mu.hat = c(1, 20, 50, 60, 65), std.err = rep(10, 5))
names(ccvx.samples$jags.samples)
ccvx_plot_fit(ccvx.samples)
}
