% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccvx_risk_profile.R
\name{ccvx_risk_profile}
\alias{ccvx_risk_profile}
\title{Plot Concavex model risk profiles}
\usage{
ccvx_risk_profile(ccvx.samples, eff.thresholds = NULL)
}
\arguments{
\item{ccvx.samples}{List containg Gibbs samples output by ccvx_fit() function}
}
\description{
Function to plot Concavex model risk profiles by effect size and dose
}
\examples{
ccvx.mod <- ccvx_build_jags()
ccvx.samples <- ccvx_fit(ccvx.mod, doses = 0:4, mu.hat = c(1, 20, 50, 60, 65), stderr = rep(10, 5))
ccvx_risk_profile(ccvx.samples)
}
